requirements.txt
python-pptx==0.6.23
pywin32==306
